change PUT method to POST with improvement
changed the PUT method to POST because its more  suitable to its job as when its run each time will insert a new record in the orderDB.

Also now the order server talks to the catalog server if they have the product or not. or if there is a quantity in the store before it does the purchase method.